Correct source folder structure
